Variety of improvements.
- Added support for dynamic log levels. Rather than putting the
log level in the [LoggerMessage] attribute, you can put it as
an argument to the logging method. This allows dynamic log levels.

- Made the LogStateHolder types classes instead of structs. This
dedicates one allocation in order to save multiple boxings of the
struct within the ILogger implementation. In addition, the generated
code ends up faster because it doesn't have to copy around a potentially
large struct.

- Fix logic that was parsing the [LoggerMessage] attribute values. it
wasn't dealing with using name: syntax to initialize the values of an
attribute.