Various improvements.
- The ILogger parameter to logging methods no longer needs to be the
first argument in the logging method. As long as it is one of the
arguments, everything will work.

- Support specifying the [LoggerMessage] attribute without an actual
message string. When this happens, an auto-generated message string
is produced which outputs the logging method parameters in JSON
format.

- Fixed bug where the LogStateHolder object was being populated with
exception and log level objects when those should have been stripped
away

- Specifying an argument to a logging method, but not supplying a
template for it in the log message is now treated as a warning instead
of an error.