A few improvements
- For instance-mode logging method, the code now binds to any single
field that implements ILogger. The code generates an error if there
are multiple matching fields. The code doesn't currently look up the
inheritance hierarchy for fields or properties that match, it certainly
could in the future if that is deemed useful.

- Remove the global option to control the field name used for instance
methods, since the field is now determined dynamically based on type.

- Made a few diagnostics into warnings since they were really just
advisory in nature.

- Cleaned up the diagnostic and resource string names for better
consistency.